% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutt.R
\name{rcutt}
\alias{rcutt}
\alias{cutt}
\alias{dcutt}
\alias{pcutt}
\alias{qcutt}
\title{The Compound Uniform-Truncated T (CUTT) Distribution for Fossil Dates}
\usage{
rcutt(n, theta, K, sd, df = NULL, tol = sqrt(.Machine$double.eps), nIter = 50)

qcutt(p, theta, K, sd, df = NULL, tol = sqrt(.Machine$double.eps), nIter = 50)

pcutt(q, theta, K, sd, df = NULL, lower.tail = TRUE, pMinus = 0)

dcutt(x, theta, K, sd, df = NULL, log = FALSE)
}
\arguments{
\item{n}{Number of observations (numeric)}

\item{theta}{Time of extinction (numeric)}

\item{K}{Numeric upper bound for fossil ages - how old fossils can be before they are ignored, for the purpose of this analysis. A sensible choice of \code{K} is
close to the age of the oldest fossil.}

\item{sd}{Measurement error standard deviations for fossils. Can be a vector. If not of length \code{n}, it is cycled through repeatedly until length \code{n}.}

\item{df}{numeric; degrees of freedom for the t-distribution used to model measurement error. Must be at least 2. Default (NULL) uses a Gaussian distribution.}

\item{tol}{Numerical tolerance (defaults to \code{sqrt(.Machine$double.eps)}}

\item{nIter}{The maximum number of iterations to use estimating random numbers before resorting to \code{uniroot} (which is slower).}

\item{p}{vector of probabilities}

\item{lower.tail}{logical; if TRUE (default), probabilities \code{p} are \eqn{P[X\leq x]}{P(X<=x)} otherwise \eqn{P[X>x]}{P(X>x)}.}

\item{pMinus}{An amount to subtract from the distribution function in \code{pcutt} (used for root-finding in \code{rcutt} and \code{qcutt}).}

\item{x, q}{vector of quantiles}

\item{log}{logical; if TRUE, densities \code{p} are given as \code{log(p)}.}
}
\value{
\code{dcutt} gives the density, \code{pcutt} gives the distribution function, \code{qnorm} 
gives the quantile function, and \code{rcutt} generates random fossil dates.

The length of the result is determined by \code{n} for \code{rcutt}, and in other cases, by the length of the first argument.
\code{sd} can also be a vector but if its not length does not match that expected by the first argument it is cycled through until the desired
length is reached, with a warning.
}
\description{
Density, distribution function, quantile function and random generation for the Compound Uniform-Truncated T (CUTT) distribution
for fossil dates with extinction time \code{theta}, oldest observable data \code{K} and measurement error \code{sd}.
}
\details{
Given an extinction time \code{theta}, an upper limit \code{K} for the possible age of a fossil
that could be included in this dataset, and measurement error standard deviations \code{sd}, our procedure then assumes that:
\itemize{
\item For a fossil of a given age, measurement error estimating its age is normally distributed with mean zero and standard deviation as provided
\item That fossil dates are uniformly distributed over the interval of allowable dates
\item All observed fossil dates are less than \code{K}
}
This leads to a distribution function of the form
\deqn{C\left[(w-\theta)\Phi\left(\frac{w-\theta}{\sigma}\right)+\sigma^2\phi\left(\frac{w-\theta}{\sigma}\right)\right]}
where \eqn{C^{-1} = (K-\theta)\Phi\left(\frac{K-\theta}{\sigma}\right)+\sigma^2\phi\left(\frac{K-\theta}{\sigma}\right)}.
}
\examples{
ages = rcutt(20, 10000, 25000, 1000) # simulate some random data
pcutt(ages, 10000, 25000, 1000) # find CDF of each value (approx uniform)
qcutt(c(0.25,0.75), 10000, 25000, 1000) # find first and third quartiles

# plot the density function
w = seq(5000,26000,length=1000)
plot(w,dcutt(w,10000,25000,1000),type="l",ylab="pdf(w)")
# compare to density if measurement error came from t(4) distribution
plot(w,dcutt(w,10000,25000,1000,df=4),type="l",ylab="pdf(w,df=4)")
}
