% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reginv_cutt.R
\name{reginv_cutt}
\alias{reginv_cutt}
\title{Confidence Interval for Extinction Time using regression inversion}
\usage{
reginv_cutt(
  ages,
  sd,
  K,
  df = Inf,
  alpha = 0.05,
  q = c(alpha/2, 0.5, 1 - alpha/2),
  paramInits = NULL,
  iterMax = 500,
  method = "rq"
)
}
\arguments{
\item{ages}{Numeric vector of fossil ages, with smaller values being more recent fossil ages.}

\item{sd}{Numeric vector of measurement error standard deviations for each fossil (listed in the same order as they appear in \code{ages}).}

\item{K}{Numeric upper bound for fossil ages - how old fossils can be before they are ignored, for the purpose of this analysis. A sensible choice of \code{K} is
close to the age of the oldest fossil.}

\item{df}{Numeric; degrees of freedom for the t-distribution used to model measurement error. Must be greater than 2. Default (Inf) uses a Gaussian distribution.}

\item{alpha}{Numeric between 0 and 1. Used to find a 100(1-\code{alpha})\% confidence interval. Defaults to 0.05 (95\% confidence intervals)}

\item{q}{Numeric vector of values between 0 and 1, specifying the quantiles at which we want to solve for extinction time. Defaults to \code{c(alpha/2,0.5,1-alpha/2)},
which gives the limits of a 100(1-\code{alpha})\% confidence interval and a point estimate obtained by solving at 0.5. If \code{q} is specified it overrides any input for \code{alpha}.}

\item{paramInits}{A numeric vector of initial values for extinction time to use in simulation. If \code{NULL} then these will be 100 values evenly distributed within 5 SEs of the estimate from \code{\link{mle_cutt}}.}

\item{iterMax}{Maximum number of simulated datasets to use to estimate extinction time (default 500).}

\item{method}{Regression method to use in estimating how MLE quantiles vary with extinction time, using a dataset of simulated MLEs and the extinction times at which they were simulated.
\code{method='rq'} (default) uses linear quantile regression, \code{method='rq2'} uses quadratic quantile regression, \code{method='wrq'} uses linear quantile regression but down-weighting 
high influence points, \code{method='prob'} uses linear probit regression on an indicator variable for whether or not simulated MLEs exceeds the observed MLE.
\code{method='lm'} uses a linear regression, which is appropriate for bias correction rather than confidence interval estimation. In this case \code{q} and \code{alpha} are ignored and only a point estimate is returned.}
}
\value{
This function returns an object of class "reginv" with the following components:

 \item{theta}{ a vector of estimated extinction times at each of a set of quantiles specified in \code{q}. (If \code{q} was not specified as input, this defaults to the lower limit for a \code{100(1-alpha)}\% confidence interval, a point estimate at \code{q=0.5} ("best estimate" of extinction time), and an upper limit for a \code{100(1-alpha)}\% confidence interval.)}
 \item{q}{ the vector of quantiles used in estimation.}
 \item{error}{ Monte Carlo standard error estimating each of these values, as estimated from the regression. }
 \item{iter}{ the number of iterations taken to estimate this value}
 \item{converged}{ whether or not this converged, at the specified tolerance}
 \item{call }{ the function call}
}
\description{
Estimates a confidence interval for extinction time using regression inversion, which finds a range of values for extinction time that
are plausible considering the maximum likelihood estimator of the most recently observed fossil. This is done assuming fossil dates come from a
a compound uniform-truncated T distribution, which accounts for sampling error (the fact that the most recent fossil date is not necessarily
the most recent time that the species was extant) and measurement error (error dating fossils).
Usually takes a few seconds to run, because it is doing a lot of computation behind the scenes.
}
\details{
Given a vector of fossil ages \code{ages} and corresponding measurement error standard deviations \code{sd}, and an upper limit \code{K} for the possible age of a fossil
that could be included in this dataset, our procedure then assumes:
\itemize{
\item That, for a fossil of a given age, measurement error estimating its age is normally distributed with mean zero and standard deviation as provided, truncated such that observed age is less than \code{K}
\item That, given a value for measurement error, fossil dates are uniformly distributed over the interval of allowable dates (which goes from estimated extinction time up until the value such that observed age is \code{K})
}
We then estimate extinction time by inversion of the maximum likelihood estimator, that is, we find the estimate of extinction time \eqn{\theta}{t}
such that if this were the true extinction time then the chance of seeing a MLE less than the one obtained from sample \code{ages} is equal to \code{q}. The probability
is estimated by regression, hence we call our procedure regression inversion. If \code{alpha} is specified this function will return three values:
the lower limit of the \code{100*(1-alpha)}\% confidence interval (solving at \code{q=alpha/2}), a point estimator for extinction time (solving at \code{q=0.5}), and an upper limit for the
confidence interval (solving at \code{q=1-alpha/2}). If a vector \code{q} is specified as input then the function solves for this vector instead. 

It is assumed that \code{ages} has been specified with smaller values representing more recent specimens, for example, \code{ages} could be specified in years before present.
If there is interest in estimating speciation or invasion time, data would only need to be reordered so that smaller values represent older specimens.
}
\examples{
ages = rcutt(20, 10000, K=25000, sd=1000) #simulating some random data

# for a point estimate together with a 95\% CI
reginv_cutt(ages=ages, sd=1000, K=25000, alpha=0.05) 

# compare to estimates using asymptotic likelihood inference, which tend to
# be narrower and have poorer coverage (they miss the true value too often
# when n or sd is small):
mle_cutt(ages=ages, sd=1000, K=25000, alpha=0.05) 
}
\seealso{
mle_cutt, rcutt
}
